feat: `getTransformStream()` export for programmatic usage
